refactor(participant-portal): extract TeamScorePanel from Home.tsx
Home.tsx held both the HomePage dashboard and an inlined TeamScorePanel
(score / rank / problem-count / solved summary card), pulling 16 imports — at
the high-coupling threshold. TeamScorePanel owned the only KeyValuePairs use and
the two portal-client API types, so extracting it to its own module drops
Home.tsx to 15 imports (off the high-coupling list) and gives each module a
single responsibility.

Relates #1418.

## Regression analysis
- Pure structural move: TeamScorePanel's body is byte-identical, only relocated
  and imported back. typecheck passes; all 274 participant-portal tests pass
  unchanged. No behavior change.

## Physical impact
- NO-OP on CFn / deployed artifacts. Frontend source reorganization only.
